Several changes: * Export new class to `body.ts` * Add documentation using JSDoc * Use `for` in `applyBaseFilters` to be able to break it
